+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 ### Added
 
 * nostr: add NIP-38 support ([reyamir])
+* nostr: add NIP-60 event kinds ([Yuki Kishimoto])
 * nostr: add NIP-62 support ([awiteb])
 * nostr: add `nip21::extract_from_text` function ([Yuki Kishimoto])
 * nostr: add `EventBuilder::allow_self_tagging` ([Yuki Kishimoto])